fine tuned som code

+++ js/product.js
@@ -1,25 +1,32 @@
-const productContainer = document.querySelector(".product-container")
-const addToBasket = document.querySelector(".basket-button")
+const productContainer = document.querySelector(".products-container")
+const addToBasket = document.querySelector(".basket-button");
+const loading = document.querySelector(".loading");
 
 const queryString = window.location.search;
 const urlParam = new URLSearchParams(queryString);
 const idValue = urlParam.get(`id`);
-console.log(idValue);
 
-let data = [];
 let dataMovie = [];
 
-fetch("https://api.noroff.dev/api/v1/square-eyes")
-    .then((response) => response.json())
-    .then((result) => {
-        data = result;
-        dataMovie = data.find(val => val.id === idValue);
+async function fetchMovieId (){
+    loading.style.display = `block`;
+    try {
+        const response = await fetch("https://api.noroff.dev/api/v1/square-eyes");
+        const result = await response.json();
+        dataMovie = result.find(val => val.id === idValue);
         displayMovie(dataMovie);
-    });
+    }
+    catch (error){
+        alert("Could not fetch data!");
+    }
+}
+
+fetchMovieId();
 
 function displayMovie (){
+    loading.style.display = `none`;
     productContainer.innerHTML +=`
-        <div>
+        <div class="product-container">
            <img class="img" src="${dataMovie.image}" alt="movie-img">
            <h1>${dataMovie.title}</h1>
            <p class="bold">price: ${dataMovie.price}
@@ -39,7 +46,10 @@ function saveLocalStorage (){
         localStorage.setItem(`data`, `[]`);
     }
     let oldData = JSON.parse(localStorage.getItem(`data`));
-    oldData.push(newData);
-    localStorage.setItem(`data`, JSON.stringify(oldData));
+    const exists = oldData.some(item => item.id === newData.id);
+    if (!exists){
+        oldData.push(newData);
+        localStorage.setItem(`data`, JSON.stringify(oldData));
+    }
 }
